fix(GxJadeLib): Use long instead of ulong/uint for GeneXus compatibility
GeneXus doesn't support ulong/uint types properly and maps them to
AnyObject. Changed GxHsmInfo properties (OperationsCount, AutoLockTimeout,
AutoLockRemaining) from ulong/uint to long, with explicit casts in the
wrapper to maintain compatibility with the underlying JadeRpc API.

+++ jade-csharp-client/src/GxJadeLib/GxJadeWrapper.cs
@@ -565,9 +565,9 @@ public static class GxJadeWrapper
                 info.TestnetRootPath = hsmInfo.TestnetRootPath ?? string.Empty;
                 info.MainnetRootPubkey = HexConverter.ToHex(hsmInfo.MainnetRootPubkey);
                 info.TestnetRootPubkey = HexConverter.ToHex(hsmInfo.TestnetRootPubkey);
-                info.OperationsCount = hsmInfo.OperationsCount;
-                info.AutoLockTimeout = hsmInfo.AutoLockTimeout;
-                info.AutoLockRemaining = hsmInfo.AutoLockRemaining ?? 0;
+                info.OperationsCount = (long)hsmInfo.OperationsCount;
+                info.AutoLockTimeout = (long)hsmInfo.AutoLockTimeout;
+                info.AutoLockRemaining = (long)(hsmInfo.AutoLockRemaining ?? 0);
             }
 
             return JadeOperationResult.Ok(connectionId);
